Correct linter error
