+++ books.json
@@ -1,1 +1,1 @@
-{"books":[{"title":"Harry Potter and the Sorcerer’s Stone","author":"JK Rowling"},{"title":"Game of Thrones","author":"George R.R. Martin"},{"title":"The Hobbit","author":"J.R.R. Tolkien"},{"title":"Alice's Adventures in Wonderland","author":"Lewis Carroll"},{"title":"Jonathan Strange","author":"Susanna Clarke"}]}
+{"books":[{"title":"Harry Potter and the Sorcerer’s Stone","author":"JK Rowling"},{"title":"Game of Thrones","author":"George R.R. Martin"},{"title":"The Hobbit","author":"J.R.R. Tolkien"},{"title":"Alice's Adventures in Wonderland","author":"Lewis Carroll"},{"title":"Jonathan Strange","author":"Susanna Clarke"},{"title":"Core Mathematics","author":"Mershark"}]}
