+++ people.json
@@ -1,1 +1,1 @@
-{"people":[{"id":1,"name":"Will","age":"10","parent_permission":true,"classroom":"10A"},{"id":1,"age":"30","name":"Ben","parent_permission":true,"specialization":"Math"}]}
+{"people":[{"id":1,"name":"Will","age":"10","parent_permission":true,"classroom":"10A"},{"id":1,"age":"30","name":"Ben","parent_permission":true,"specialization":"Math"},{"id":2,"name":"sdf","age":"12","parent_permission":true,"classroom":"asd"},{"id":2,"age":"12","name":"asf","parent_permission":true,"specialization":"math"},{"id":3,"name":"mershark","age":"12","parent_permission":true,"classroom":"one"}]}
