+++ rentals.json
@@ -1,1 +1,1 @@
-{"rentals":[{"date":"2023/10/10","book":{"title":"Harry Potter and the Sorcerer’s Stone","author":"JK Rowling"},"person":{"id":1,"name":"Will","age":"10","parent_permission":true,"classroom":"10A"}}]}
+{"rentals":[{"date":"1999/02/02","book":{"title":"Harry Potter and the Sorcerer’s Stone","author":"JK Rowling"},"person":{"id":1,"name":"Will","age":"10","parent_permission":true,"classroom":"10A"}}]}
